requirements.txt
Django==1.4.3
Pillow==1.7.8
South==0.7.6
Whoosh==2.4.1
amqp==1.0.6
anyjson==0.3.3
billiard==2.7.3.19
celery==3.0.13
django-activity-stream==0.4.4
django-appconf==0.5
django-celery==3.0.11
django-compressor==1.2
django-countries==1.5
django-debug-toolbar==0.9.4
django-haystack==1.2.7
django-imagekit==2.0.2
django-registration==0.8
django-social-auth==0.7.18
djangorestframework==2.1.16
httplib2==0.7.7
kombu==2.5.4
oauth2==1.5.211
psycopg2==2.4.6
python-dateutil==1.5
python-openid==2.2.5
pytz==2012j
wsgiref==0.1.2
